Rename `AbstractHandlerMethod` to `AbstractReceptor`
