+++ server/src/test/java/io/spine/server/model/given/method/OneParamMethod.java
@@ -27,7 +27,7 @@
 package io.spine.server.model.given.method;
 
 import io.spine.base.EventMessage;
-import io.spine.server.model.AbstractHandlerMethod;
+import io.spine.server.model.AbstractReceptor;
 import io.spine.server.model.ParameterSpec;
 import io.spine.server.model.VoidMethod;
 import io.spine.server.type.EmptyClass;
@@ -37,11 +37,11 @@ import io.spine.server.type.EventEnvelope;
 import java.lang.reflect.Method;
 
 public class OneParamMethod
-        extends AbstractHandlerMethod<Object,
-                                      EventMessage,
-                                      EventClass,
-                                      EventEnvelope,
-                                      EmptyClass>
+        extends AbstractReceptor<Object,
+                                 EventMessage,
+                                 EventClass,
+                                 EventEnvelope,
+                                 EmptyClass>
         implements VoidMethod<Object, EventClass, EventEnvelope> {
 
     public OneParamMethod(Method method, ParameterSpec<EventEnvelope> parameterSpec) {
